# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Video Processing API",
    "description": "API para procesamiento de video optimizada para creación de contenido",
    "version": "1.0.0",
    "contact": {
      "name": "VideoAPI Support",
      "email": "support@sofe.site"
    }
  },
  "servers": [
    {
      "url": "https://videoapi.sofe.site",
      "description": "Servidor de producción"
    },
    {
      "url": "http://localhost:8080",
      "description": "Servidor de desarrollo"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "error": {
            "type": "string",
            "example": "Internal Server Error"
          },
          "message": {
            "type": "string",
            "example": "Ocurrió un error inesperado en el servidor"
          }
        }
      },
      "JobResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "queued"
          },
          "job_id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "message": {
            "type": "string",
            "example": "Task has been queued for processing"
          }
        }
      },
      "JobStatus": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "job_id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "job_status": {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "example": "completed"
              },
              "created_at": {
                "type": "number",
                "example": 1641377400.34
              },
              "completed_at": {
                "type": "number",
                "example": 1641377430.12
              },
              "result": {
                "type": "string",
                "example": "https://videoapi.sofe.site/storage/abc123.mp4"
              }
            }
          }
        }
      },
...
      "CleanupResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "files_removed": {
            "type": "integer",
            "example": 15
          },
          "space_freed_mb": {
            "type": "number",
            "example": 256.34
          }
        }
      },
      "MemeOverlayRequest": {
        "type": "object",
        "required": ["video_url", "meme_url"],
        "properties": {
          "video_url": {
            "type": "string",
            "format": "uri",
            "example": "https://example.com/video.mp4"
          },
          "meme_url": {
            "type": "string",
            "format": "uri",
            "example": "https://example.com/meme.png"
          },
          "position": {
            "type": "string",
            "enum": ["top", "bottom", "left", "right", "top_left", "top_right", "bottom_left", "bottom_right", "center"],
            "default": "bottom",
            "example": "bottom"
          },
          "scale": {
            "type": "number",
            "minimum": 0.1,
            "maximum": 1.0,
            "default": 0.3,
            "example": 0.3
          },
          "webhook_url": {
            "type": "string",
            "format": "uri",
            "example": "https://webhook.example.com/callback"
          },
          "id": {
            "type": "string",
            "example": "custom-job-id-123"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/maintenance/cleanup": {
      "post": {
        "summary": "Limpieza manual",
        "description": "Inicia manualmente la limpieza de archivos antiguos",
        "tags": ["Maintenance"],
        "responses": {
          "200": {
            "description": "Resultado de la limpieza",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "summary": "Estado de un trabajo",
        "description": "Consulta el estado de un trabajo específico",
        "tags": ["Jobs"],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del trabajo a consultar"
          }
        ],
        "responses": {
          "200": {
            "description": "Estado del trabajo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Trabajo no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/storage/{filename}": {
      "get": {
        "summary": "Descarga de archivo",
        "description": "Descarga un archivo del almacenamiento",
        "security": [],
        "tags": ["Storage"],
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Nombre del archivo a descargar"
          }
        ],
        "responses": {
          "200": {
            "description": "Archivo",
            "content": {
              "video/mp4": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "audio/mp3": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "403": {
            "description": "Acceso no autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Archivo no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/video/meme_overlay": {
      "post": {
        "summary": "Superponer meme en video",
        "description": "Superpone una imagen de meme sobre un video",
        "tags": ["Video"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemeOverlayRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operación completada con éxito (procesamiento inmediato)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "job_id": {
                      "type": "string",
                      "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "result": {
                      "type": "string",
                      "example": "https://videoapi.sofe.site/storage/abc123.mp4"
                    },
                    "endpoint": {
                      "type": "string",
                      "example": "/v1/video/meme_overlay"
                    }
                  }
                }
              }
            }
          },
          "202": {
            "description": "Trabajo encolado para procesamiento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/video/caption_video": {
      "post": {
        "summary": "Añadir subtítulos a video",
        "description": "Añade subtítulos a un video desde un archivo SRT, VTT, etc.",
        "tags": ["Video"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptionVideoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operación completada con éxito (procesamiento inmediato)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "job_id": {
                      "type": "string",
                      "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "result": {
                      "type": "string",
                      "example": "https://videoapi.sofe.site/storage/abc123.mp4"
                    },
                    "endpoint": {
                      "type": "string",
                      "example": "/v1/video/caption_video"
                    }
                  }
                }
              }
            }
          },
          "202": {
            "description": "Trabajo encolado para procesamiento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/video/concatenate": {
      "post": {
        "summary": "Concatenar videos",
        "description": "Concatena múltiples videos en uno solo",
        "tags": ["Video"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcatenateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operación completada con éxito (procesamiento inmediato)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "job_id": {
                      "type": "string",
                      "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "result": {
                      "type": "string",
                      "example": "https://videoapi.sofe.site/storage/abc123.mp4"
                    },
                    "endpoint": {
                      "type": "string",
                      "example": "/v1/video/concatenate"
                    }
                  }
                }
              }
            }
          },
          "202": {
            "description": "Trabajo encolado para procesamiento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}